requirements.txt
python-dotenv
sqlalchemy
psycopg2-binary
requests
python-docx
pymupdf
fpdf
google-generativeai